update unsgined and remove some diff

+++ src/flags.cc
@@ -15,7 +15,7 @@ std::string FLAGS_master_addr = "localhost:" + FLAGS_master_port;
 int FLAGS_task_retry_times = 3;
 int FLAGS_task_deploy_timeout = 20;
 int FLAGS_agent_keepalive_timeout = 20;
-unsigned long FLAGS_master_max_len_sched_task_list = 30;
+int FLAGS_master_max_len_sched_task_list = 30;
 
 std::string FLAGS_agent_work_dir = "/tmp";
 std::string FLAGS_container = "cmd";// cmd or cgroup
